+++ data/trie/trieStorageManager_test.go
@@ -1,13 +1,10 @@
 package trie
 
 import (
-	"errors"
 	"io/ioutil"
 	"os"
 	"path"
-	"runtime"
 	"strconv"
-	"strings"
 	"sync"
 	"testing"
 	"time"
@@ -21,10 +18,15 @@ import (
 )
 
 const (
-	snapshotDelay                 = time.Second
 	checkpointHashesHolderMaxSize = 10000000
 	hashSize                      = 32
 )
+
+func waitForOperationToComplete(tsm data.StorageManager) {
+	for tsm.IsPruningBlocked() {
+		time.Sleep(10 * time.Millisecond)
+	}
+}
 
 func getNewTrieStorageManagerArgs() NewTrieStorageManagerArgs {
 	return NewTrieStorageManagerArgs{
@@ -120,7 +122,7 @@ func TestNewTrieStorageManagerWithExistingSnapshot(t *testing.T) {
 	_ = tr.Commit()
 	rootHash, _ := tr.RootHash()
 	trieStorage.TakeSnapshot(rootHash, true, nil)
-	time.Sleep(snapshotDelay)
+	waitForOperationToComplete(trieStorage)
 
 	trieStorage.storageOperationMutex.Lock()
 	_ = trieStorage.snapshots[0].Close()
@@ -173,7 +175,7 @@ func TestNewTrieStorageManagerLoadsSnapshotsInOrder(t *testing.T) {
 	_ = tr.Commit()
 	rootHash, _ := tr.RootHash()
 	trieStorage.TakeSnapshot(rootHash, true, nil)
-	time.Sleep(snapshotDelay)
+	waitForOperationToComplete(trieStorage)
 
 	numSnapshots := 10
 	for i := 0; i < numSnapshots; i++ {
@@ -181,7 +183,7 @@ func TestNewTrieStorageManagerLoadsSnapshotsInOrder(t *testing.T) {
 		_ = tr.Commit()
 		rootHash, _ = tr.RootHash()
 		trieStorage.TakeSnapshot(rootHash, true, nil)
-		time.Sleep(snapshotDelay)
+		waitForOperationToComplete(trieStorage)
 	}
 
 	trieStorage.storageOperationMutex.Lock()
@@ -226,7 +228,7 @@ func TestRecreateTrieFromSnapshotDb(t *testing.T) {
 	storageManager := tr.GetStorageManager()
 	rootHash, _ := tr.RootHash()
 	storageManager.TakeSnapshot(rootHash, true, nil)
-	time.Sleep(snapshotDelay)
+	waitForOperationToComplete(storageManager)
 
 	err := storageManager.Database().Remove(rootHash)
 	assert.Nil(t, err)
@@ -258,7 +260,7 @@ func TestEachSnapshotCreatesOwnDatabase(t *testing.T) {
 		_ = tr.Update(testVal.key, testVal.value)
 		_ = tr.Commit()
 		trieStorage.TakeSnapshot(tr.root.getHash(), true, nil)
-		time.Sleep(snapshotDelay)
+		waitForOperationToComplete(trieStorage)
 
 		trieStorage.storageOperationMutex.Lock()
 		snapshotId := strconv.Itoa(trieStorage.snapshotId - 1)
@@ -293,7 +295,7 @@ func TestDeleteOldSnapshots(t *testing.T) {
 		_ = tr.Commit()
 		trieStorage.TakeSnapshot(tr.root.getHash(), true, nil)
 	}
-	time.Sleep(snapshotDelay)
+	waitForOperationToComplete(trieStorage)
 
 	snapshots, _ := ioutil.ReadDir(trieStorage.snapshotDbCfg.FilePath)
 	assert.Equal(t, 2, len(snapshots))
@@ -317,7 +319,7 @@ func TestTrieCheckpoint(t *testing.T) {
 
 	_ = tr.Commit()
 	trieStorage.TakeSnapshot(tr.root.getHash(), true, nil)
-	time.Sleep(snapshotDelay)
+	waitForOperationToComplete(trieStorage)
 
 	_ = tr.Update([]byte("doge"), []byte("reindeer"))
 	newHashes, _ := tr.GetDirtyHashes()
@@ -347,7 +349,7 @@ func TestTrieCheckpoint(t *testing.T) {
 	assert.Nil(t, val)
 
 	trieStorage.SetCheckpoint(tr.root.getHash(), nil)
-	time.Sleep(snapshotDelay)
+	waitForOperationToComplete(trieStorage)
 
 	val, err = snapshotTrie.Get([]byte("doge"))
 	assert.Nil(t, err)
@@ -366,7 +368,7 @@ func TestTrieCheckpointWithNoSnapshotCreatesSnapshot(t *testing.T) {
 
 	_ = tr.Commit()
 	trieStorage.SetCheckpoint(tr.root.getHash(), nil)
-	time.Sleep(snapshotDelay)
+	waitForOperationToComplete(trieStorage)
 
 	trieStorage.storageOperationMutex.Lock()
 	assert.Equal(t, 1, len(trieStorage.snapshots))
@@ -383,7 +385,7 @@ func TestTrieSnapshottingAndCheckpointConcurrently(t *testing.T) {
 	_ = tr.Commit()
 
 	trieStorage.TakeSnapshot(tr.root.getHash(), true, nil)
-	time.Sleep(snapshotDelay)
+	waitForOperationToComplete(trieStorage)
 
 	numSnapshots := 5
 	numCheckpoints := 5
@@ -424,7 +426,7 @@ func TestTrieSnapshottingAndCheckpointConcurrently(t *testing.T) {
 
 	snapshotWg.Wait()
 	checkpointWg.Wait()
-	time.Sleep(snapshotDelay * 3)
+	waitForOperationToComplete(trieStorage)
 
 	trieStorage.storageOperationMutex.Lock()
 	assert.Equal(t, totalNumSnapshot, trieStorage.snapshotId)
@@ -452,14 +454,14 @@ func TestIsPresentInLastSnapshotDb(t *testing.T) {
 	_ = tr.Commit()
 	rootHash1, _ := tr.RootHash()
 	trieStorage.TakeSnapshot(rootHash1, true, nil)
-	time.Sleep(snapshotDelay)
+	waitForOperationToComplete(trieStorage)
 
 	_ = tr.Update([]byte("dog"), []byte("puppy"))
 
 	_ = tr.Commit()
 	rootHash2, _ := tr.RootHash()
 	trieStorage.TakeSnapshot(rootHash2, true, nil)
-	time.Sleep(snapshotDelay)
+	waitForOperationToComplete(trieStorage)
 
 	trieStorage.storageOperationMutex.Lock()
 	val, err := trieStorage.snapshots[0].Get(rootHash2)
@@ -483,14 +485,14 @@ func TestTrieSnapshotChecksOnlyLastSnapshotDbForTheHash(t *testing.T) {
 	_ = tr.Commit()
 	rootHash1, _ := tr.RootHash()
 	trieStorage.TakeSnapshot(rootHash1, true, nil)
-	time.Sleep(snapshotDelay)
+	waitForOperationToComplete(trieStorage)
 
 	_ = tr.Update([]byte("dog"), []byte("puppy"))
 
 	_ = tr.Commit()
 	rootHash2, _ := tr.RootHash()
 	trieStorage.TakeSnapshot(rootHash2, true, nil)
-	time.Sleep(snapshotDelay)
+	waitForOperationToComplete(trieStorage)
 
 	trieStorage.storageOperationMutex.Lock()
 	val, err := trieStorage.snapshots[0].Get(rootHash1)
@@ -514,14 +516,14 @@ func TestShouldNotRemoveSnapshotDbIfItIsStillInUse(t *testing.T) {
 	_ = tr.Commit()
 	rootHash1, _ := tr.RootHash()
 	trieStorage.TakeSnapshot(rootHash1, true, nil)
-	time.Sleep(snapshotDelay)
+	waitForOperationToComplete(trieStorage)
 
 	_ = tr.Update([]byte("dog"), []byte("puppy"))
 
 	_ = tr.Commit()
 	rootHash2, _ := tr.RootHash()
 	trieStorage.TakeSnapshot(rootHash2, true, nil)
-	time.Sleep(snapshotDelay)
+	waitForOperationToComplete(trieStorage)
 
 	db := trieStorage.GetSnapshotThatContainsHash(rootHash1)
 
@@ -530,7 +532,7 @@ func TestShouldNotRemoveSnapshotDbIfItIsStillInUse(t *testing.T) {
 	_ = tr.Commit()
 	rootHash3, _ := tr.RootHash()
 	trieStorage.TakeSnapshot(rootHash3, true, nil)
-	time.Sleep(snapshotDelay)
+	waitForOperationToComplete(trieStorage)
 
 	val, err := db.Get(rootHash1)
 	assert.Nil(t, err)
@@ -556,7 +558,7 @@ func TestShouldNotRemoveSnapshotDbsIfKeepSnapshotsTrue(t *testing.T) {
 		_ = tr.Commit()
 		rootHash, _ := tr.RootHash()
 		trieStorage.TakeSnapshot(rootHash, true, nil)
-		time.Sleep(snapshotDelay)
+		waitForOperationToComplete(trieStorage)
 	}
 
 	for i := 0; i < nrOfSnapshots; i++ {
@@ -584,7 +586,7 @@ func TestShouldRemoveSnapshotDbsIfKeepSnapshotsFalse(t *testing.T) {
 		_ = tr.Commit()
 		rootHash, _ := tr.RootHash()
 		trieStorage.TakeSnapshot(rootHash, true, nil)
-		time.Sleep(snapshotDelay)
+		waitForOperationToComplete(trieStorage)
 	}
 
 	for i := 0; i < nrOfSnapshots-int(trieStorage.maxSnapshots); i++ {
@@ -615,14 +617,14 @@ func TestShouldNotDisconnectSnapshotDbIfItIsStillInUse(t *testing.T) {
 	_ = tr.Commit()
 	rootHash1, _ := tr.RootHash()
 	trieStorage.TakeSnapshot(rootHash1, true, nil)
-	time.Sleep(snapshotDelay)
+	waitForOperationToComplete(trieStorage)
 
 	_ = tr.Update([]byte("dog"), []byte("puppy"))
 
 	_ = tr.Commit()
 	rootHash2, _ := tr.RootHash()
 	trieStorage.TakeSnapshot(rootHash2, true, nil)
-	time.Sleep(snapshotDelay)
+	waitForOperationToComplete(trieStorage)
 
 	db := trieStorage.GetSnapshotThatContainsHash(rootHash1)
 
@@ -631,7 +633,7 @@ func TestShouldNotDisconnectSnapshotDbIfItIsStillInUse(t *testing.T) {
 	_ = tr.Commit()
 	rootHash3, _ := tr.RootHash()
 	trieStorage.TakeSnapshot(rootHash3, true, nil)
-	time.Sleep(snapshotDelay)
+	waitForOperationToComplete(trieStorage)
 
 	val, err := db.Get(rootHash1)
 	assert.Nil(t, err)
@@ -710,50 +712,4 @@ func TestTrieStorageManager_Remove(t *testing.T) {
 	assert.NotNil(t, err)
 	ok = args.CheckpointHashesHolder.ShouldCommit(key)
 	assert.False(t, ok)
-}
-
-func TestTrieStorageManager_Close(t *testing.T) {
-	closeCalled := false
-	args := getNewTrieStorageManagerArgs()
-	args.DB = &mock.StorerStub{
-		CloseCalled: func() error {
-			closeCalled = true
-			return nil
-		},
-	}
-	numBefore := runtime.NumGoroutine()
-	ts, _ := NewTrieStorageManager(args)
-	numGoRoutines := runtime.NumGoroutine()
-	assert.Equal(t, 1, numGoRoutines-numBefore)
-
-	err := ts.Close()
-	assert.Nil(t, err)
-	time.Sleep(time.Second)
-	numAfterClose := runtime.NumGoroutine()
-	assert.Equal(t, 0, numAfterClose-numBefore)
-	assert.True(t, closeCalled)
-}
-
-func TestTrieStorageManager_CloseErr(t *testing.T) {
-	closeCalled := false
-	closeErr := errors.New("close error")
-	args := getNewTrieStorageManagerArgs()
-	args.DB = &mock.StorerStub{
-		CloseCalled: func() error {
-			closeCalled = true
-			return closeErr
-		},
-	}
-	numBefore := runtime.NumGoroutine()
-	ts, _ := NewTrieStorageManager(args)
-	numGoRoutines := runtime.NumGoroutine()
-	assert.Equal(t, 1, numGoRoutines-numBefore)
-
-	err := ts.Close()
-	assert.NotNil(t, err)
-	assert.True(t, strings.Contains(err.Error(), closeErr.Error()))
-	time.Sleep(time.Second)
-	numAfterClose := runtime.NumGoroutine()
-	assert.Equal(t, 0, numAfterClose-numBefore)
-	assert.True(t, closeCalled)
 }
